minor refactoring of code that handles track coordinates
